+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.5] - 2020-10-03
+### Changed
+- Made refresh button interactive, and reloads the page on click.
+
+### Removed
+- Ability to show all questions for a prompt (keep it simple!)
+
+### Added
+- Version number in colophon.
+
 ## [0.0.4] - 2020-09-28
 ### Changed
 - Moved refresh instruction out of the colophon, where it was hidden.